requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
coverage==7.4.0
Django==5.0
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.13.1
idna==3.6
inflection==0.5.1
packaging==23.2
Pillow==10.1.0
platformdirs==4.1.0
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
stripe==7.12.0
typing_extensions==4.9.0
uritemplate==4.1.1
uritools==4.0.2
urlextract==1.8.0
urllib3==2.1.0
